# apps/api-gateway/doc/openapi.yml
...
openapi: 3.0.3
info:
  title: Movie Service API
  version: 1.0.0
  description: Quản lý thông tin website đặt vé xem phim
servers:
  - url: /api/{version}
    variables:
      version:
        default: v1
        enum:
          - v1
          - v2
...
paths:
...
  /movies:
    get:
      tags: [Movies]
      summary: Lấy danh sách phim
      parameters:
        - $ref: '#/components/parameters/PageParameter'
        - $ref: '#/components/parameters/LimitParameter'
        - $ref: '#/components/parameters/SortByParameter'
        - $ref: '#/components/parameters/SortOrderParameter'
        - name: status
          in: query
          schema:
            enum: [now_showing, upcoming]

      responses:
        '200':
          description: Danh sách phim
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MovieSummary'
                  meta:
                    $ref: '#/components/schemas/ListMeta'
        '400':
          description: Lỗi request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    post:
      tags: [Movies]
      summary: Tạo phim mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMovieDto'
      responses:
        '201':
          description: Phim được tạo thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MovieDetail'
        '400':
          description: Lỗi request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /movies/{movieId}:
    get:
      tags: [Movies]
      summary: Lấy chi tiết phim
      parameters:
        - name: movieId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Chi tiết phim
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MovieDetail'
        '400':
          description: Lỗi request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    put:
      tags: [Movies]
      summary: Cập nhật thông tin phim
      parameters:
        - name: movieId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMovieDto'
      responses:
        '200':
          description: Phim được cập nhật thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MovieDetail'
        '400':
          description: Lỗi request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    delete:
      tags: [Movies]
      summary: Xóa phim
      parameters:
        - name: movieId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Phim được xóa thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessResponse'
        '400':
          description: Lỗi request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /movies/{movieId}/reviews:
    get:
      tags: [Movies]
      summary: Lấy chi tiết bình luận phim
      parameters:
        - name: movieId
          in: path
          schema:
            type: string
            format: uuid
          required: true
        - $ref: '#/components/parameters/PageParameter'
        - $ref: '#/components/parameters/LimitParameter'
        - $ref: '#/components/parameters/SortByParameter'
        - $ref: '#/components/parameters/SortOrderParameter'
        - name: movieId
          in: query
          schema:
            type: string
            format: uuid
        - name: rating
          in: query
          schema:
            type: integer
        - name: userId
          in: query
          schema:
            type: string

      responses:
        '200':
          description: Chi tiết phim
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ReviewResponse'
        '400':
          description: Lỗi request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    post:
      tags: [Movies]
      summary: Tạo bình luận mới cho phim
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewDto'
      responses:
        '201':
          description: Phim được tạo thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReviewResponse'
        '400':
          description: Lỗi request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /movies/{movieId}/reviews/{reviewId}:
    put:
      tags: [Movies]
      summary: Sửa bình luận cho phim
      parameters:
        - name: movieId
          in: path
          schema:
            type: string
            format: uuid
          required: true
        - name: reviewId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReviewDto'
      responses:
        '201':
          description: Phim được tạo thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReviewResponse'
        '400':
          description: Lỗi request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /movies/{movieId}/releases:
    get:
      tags: [Movies]
      summary: Lấy chi tiết lịch phát hành phim
      parameters:
        - name: movieId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Chi tiết lịch phát hành phim
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/MovieRelease'
        '400':
          description: Lỗi request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /movie-releases:
    post:
      tags: [Movie Releases]
      summary: Tạo lịch phát hành phim mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMovieReleaseDto'
      responses:
        '201':
          description: Lịch phát hành phim được tạo thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MovieRelease'
        '400':
          description: Lỗi request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /movie-releases/{movieReleaseId}:
    put:
      tags: [Movie Releases]
      summary: Cập nhật lịch phát hành phim
      parameters:
        - name: movieReleaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMovieReleaseDto'
      responses:
        '200':
          description: Lịch phát hành phim được cập nhật thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessResponse'
        '400':
          description: Lỗi request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    delete:
      tags: [Movie Releases]
      summary: Xóa lịch phát hành phim
      parameters:
        - name: movieReleaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lịch phát hành phim được xóa thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessResponse'
        '400':
          description: Lỗi request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  /staffs:
    get:
      tags: [Staffs]
      summary: Lấy danh sách nhân viên
      parameters:
        # Pagination
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1

        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            default: 10

        # Sorting
        - in: query
          name: sortBy
          schema:
            type: string
            example: createdAt

        - in: query
          name: sortOrder
          schema:
            type: string
            enum: [asc, desc]
            default: asc

        # Filters
        - in: query
          name: cinemaId
          schema:
            type: string
            format: uuid

        - in: query
          name: fullName
          schema:
            type: string

        - in: query
          name: gender
          schema:
            type: string
            enum: [MALE, FEMALE, OTHER]

        - in: query
          name: dob
          schema:
            type: string
            format: date

        - in: query
          name: position
          schema:
            type: string
            enum:
              - CINEMA_MANAGER
              - ASSISTANT_MANAGER
              - TICKET_CLERK
              - CONCESSION_STAFF
              - USHER
              - PROJECTIONIST
              - CLEANER
              - SECURITY

        - in: query
          name: status
          schema:
            type: string
            enum: [ACTIVE, INACTIVE]

        - in: query
          name: workType
          schema:
            type: string
            enum: [FULL_TIME, PART_TIME, CONTRACT]

        - in: query
          name: shiftType
          schema:
            type: string
            enum: [MORNING, AFTERNOON, NIGHT]

      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/StaffResponse'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
                  message:
                    type: string
    post:
      tags: [Staffs]
      summary: Create staff
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStaffDto'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/StaffResponse'
                  message:
                    type: string
  /staffs/{id}:
    get:
      tags: [Staffs]
      summary: Lấy chi tiết nhân viên
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/StaffResponse'
                  message:
                    type: string
    put:
      tags: [Staffs]
      summary: Update staff
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStaffDto'
      responses:
        200:
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/StaffResponse'
                  message:
                    type: string
...
components:
  schemas:
    ListMeta:
      type: object
      properties:
        page: { type: integer, default: 1 }
        limit: { type: integer, default: 20 }
        totalRecords: { type: integer }
        totalPages: { type: integer }
        hasPrev: { type: boolean }
        hasNext: { type: boolean }
    ErrorItem:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
    ApiErrorResponse:
      type: object
      required: [success, message, errors, timestamp, path, statusCode]
      properties:
        success: { type: boolean, enum: [false] }
        message: { type: string }
        errors:
          type: array
          nullable: true
          items: { $ref: '#/components/schemas/ErrorItem' }
        timestamp: { type: string, format: date-time }
        path: { type: string }
        statusCode: { type: integer }
    ApiSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          nullable: true
        data:
          nullable: true
        meta:
          type: object
          nullable: true
        timestamp:
          type: string
          format: date-time
        path: { type: string }
      required: [success, message, timestamp]
    MovieSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        posterUrl:
          type: string
          format: uri
        backdropUrl:
          type: string
        runtime:
          type: integer
          description: Duration in minutes
        ageRating:
          type: string
          enum: ['P', 'K', 'T13', 'T16', 'T18', 'C']
        productionCountry:
          type: string
        languageType:
          type: string
          enum: ['ORIGINAL', 'SUBTITLE', 'DUBBED']
        averageRating:
          type: number
        reviewCount:
          type: number
    MovieDetail:
      allOf:
        - $ref: '#/components/schemas/MovieSummary'
        - type: object
          properties:
            originalTitle:
              type: string
            overview:
              type: string
            cast:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  profileUrl:
                    type: string
            trailerUrl:
              type: string
            originalLanguage:
              type: string
            spokenLanguages:
              type: array
              items:
                type: number
            genre:
              type: object
              $ref: '#/components/schemas/GenreResponse'
    GenreResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    MovieCast:
      type: object
      properties:
        name:
          type: string
          example: Patrick Wilson
        profileUrl:
          type: string
          nullable: true
          example: https://example.com/actor1.jpg
      required:
        - name
    CreateMovieDto:
      type: object
      properties:
        title:
          type: string
          example: Test 1
        overview:
          type: string
          example: Hai nhà điều tra hiện tượng siêu nhiên...
        originalTitle:
          type: string
          example: 'The Conjuring: Last Rites'
        posterUrl:
          type: string
          format: uri
          example: https://image.tmdb.org/t/p/w500/sjvW985erG7NSKJScFNhLbjhyma.jpg
        trailerUrl:
          type: string
          format: uri
          nullable: true
          example: https://www.youtube.com/watch?v=3iBhELud1-I
        backdropUrl:
          type: string
          format: uri
          nullable: true
          example: https://image.tmdb.org/t/p/w500/kHZaX0vuhZdbuq0WKU3BpA9WIQ0.jpg
        runtime:
          type: number
          example: 135
        releaseDate:
          type: string
          format: date
          example: 2025-09-03
        ageRating:
          type: string
          enum: [G, PG, PG-13, R, NC-17, P]
          example: P
        originalLanguage:
          type: string
          example: en
        spokenLanguages:
          type: array
          items:
            type: string
            example: en
        languageType:
          type: string
          enum: [ORIGINAL, DUBBED]
          example: ORIGINAL
        productionCountry:
          type: string
          example: US
        director:
          type: string
          nullable: true
          example: Michael Chaves
        cast:
          type: array
          items:
            $ref: '#/components/schemas/MovieCast'
        genreIds:
          type: array
          items:
            type: string
            format: uuid
            example: 4a507691-86fb-4daf-a1e4-ece6f91c4571
      required:
        - title
        - overview
        - posterUrl
        - runtime
        - releaseDate
        - ageRating
        - originalLanguage
        - spokenLanguages
        - languageType
        - productionCountry
        - genreIds
    UpdateMovieDto:
      type: object
      properties:
        title:
          type: string
          example: Test 1
        overview:
          type: string
          example: Hai nhà điều tra hiện tượng siêu nhiên...
        originalTitle:
          type: string
          example: 'The Conjuring: Last Rites'
        posterUrl:
          type: string
          format: uri
          example: https://image.tmdb.org/t/p/w500/sjvW985erG7NSKJScFNhLbjhyma.jpg
        trailerUrl:
          type: string
          format: uri
          nullable: true
          example: https://www.youtube.com/watch?v=3iBhELud1-I
        backdropUrl:
          type: string
          format: uri
          nullable: true
          example: https://image.tmdb.org/t/p/w500/kHZaX0vuhZdbuq0WKU3BpA9WIQ0.jpg
        runtime:
          type: number
          example: 135
        releaseDate:
          type: string
          format: date
          example: 2025-09-03
        ageRating:
          type: string
          enum: [G, PG, PG-13, R, NC-17, P]
          example: P
        originalLanguage:
          type: string
          example: en
        spokenLanguages:
          type: array
          items:
            type: string
            example: en
        languageType:
          type: string
          enum: [ORIGINAL, DUBBED]
          example: ORIGINAL
        productionCountry:
          type: string
          example: US
        director:
          type: string
          nullable: true
          example: Michael Chaves
        cast:
          type: array
          items:
            $ref: '#/components/schemas/MovieCast'
        genreIds:
          type: array
          items:
            type: string
            format: uuid
            example: 4a507691-86fb-4daf-a1e4-ece6f91c4571
    CreateMovieReleaseDto:
      type: object
      properties:
        movieId:
          type: string
          format: uuid
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
        startDate:
          type: string
          format: date
          example: 2025-09-10
        endDate:
          type: string
          format: date
          example: 2025-09-10
        note:
          type: string
          example: Phát hành dịp lễ quốc khánh
      required:
        - movieId
        - startDate
        - endDate
        - note
    UpdateMovieReleaseDto:
      type: object
      properties:
        movieId:
          type: string
          format: uuid
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
        startDate:
          type: string
          format: date
          example: 2025-09-10
        endDate:
          type: string
          format: date
          example: 2025-09-10
        note:
          type: string
          example: Phát hành dịp lễ quốc khánh
    CreateReviewDto:
      type: object
      properties:
        movieId:
          type: string
          format: uuid
        userId:
          type: string
        rating:
          type: integer
        content:
          type: string
      required:
        - movieId
        - userId
        - rating
        - content
    UpdateReviewDto:
      type: object
      properties:
        movieId:
          type: string
          format: uuid
        userId:
          type: string
        rating:
          type: integer
        content:
          type: string
    ReviewResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        movieId:
          type: string
          format: uuid
        userId:
          type: string
        rating:
          type: integer
        content:
          type: string
        createdAt:
          type: string
          format: date
        updatedAt:
          type: string
          format: date
    MovieRelease:
      type: object
      properties:
        id:
          type: string
          format: uuid
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        note:
          type: string

    # ----------------------
    # Staff
    # ----------------------
    CreateStaffDto:
      type: object
      required:
        - cinemaId
        - fullName
        - email
        - phone
        - gender
        - dob
        - position
        - status
        - workType
        - shiftType
        - salary
        - hireDate
      properties:
        cinemaId:
          type: string
          format: uuid

        fullName:
          type: string
          minLength: 1
          example: 'Nguyễn Văn A'

        email:
          type: string
          format: email
          example: 'nguyenvana@example.com'

        phone:
          type: string
          minLength: 9
          example: '0987654321'

        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]

        dob:
          type: string
          format: date
          example: '1998-06-15'

        position:
          type: string
          enum:
            - CINEMA_MANAGER
            - ASSISTANT_MANAGER
            - TICKET_CLERK
            - CONCESSION_STAFF
            - USHER
            - PROJECTIONIST
            - CLEANER
            - SECURITY

        status:
          type: string
          enum: [ACTIVE, INACTIVE]

        workType:
          type: string
          enum: [FULL_TIME, PART_TIME, CONTRACT]

        shiftType:
          type: string
          enum: [MORNING, AFTERNOON, NIGHT]

        salary:
          type: number
          format: double
          example: 8000000

        hireDate:
          type: string
          format: date
          example: '2024-01-10'
...
    StaffResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid

        cinemaId:
          type: string
          format: uuid

        fullName:
          type: string

        email:
          type: string
          format: email

        phone:
          type: string

        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]

        dob:
          type: string
          format: date

        position:
          type: string
          enum:
            - CINEMA_MANAGER
            - ASSISTANT_MANAGER
            - TICKET_CLERK
            - CONCESSION_STAFF
            - USHER
            - PROJECTIONIST
            - CLEANER
            - SECURITY

        status:
          type: string
          enum: [ACTIVE, INACTIVE]

        workType:
          type: string
          enum: [FULL_TIME, PART_TIME, CONTRACT]

        shiftType:
          type: string
          enum: [MORNING, AFTERNOON, NIGHT]

        salary:
          type: number
          format: double

        hireDate:
          type: string
          format: date

    # ----------------------
    # CINEMA
    # ----------------------
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: number
        limit:
          type: number
        totalRecords:
          type: number
        totalPages:
          type: number
        hasPrev:
          type: boolean
        hasNext:
          type: boolean
...
  parameters:
    PageParameter:
      name: page
      in: query
      description: Page number (starts from 1)
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    LimitParameter:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
    SortByParameter:
      name: sortBy
      in: query
      description: Field to sort by
      schema:
        type: string
        default: createdAt
      example: id
    SortOrderParameter:
      name: sortOrder
      in: query
      description: Sort order
      schema:
        type: string
        enum: [asc, desc]
        default: desc
      example: asc
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT